[TASK] Proof of concept for a wireframe component
The default view of the page layout module can be easily refactored
by using the form engine. The decomposition is in such a way designed that,
using other relations than `pages` and `tt_content` should be possible,
it would be easily to extend and reuse for vendor solutions, and the overall
code coverage in this area could be significantly increased.

Neither the side by side view of content elements in various languages nor
the concrete integration of the refactored default view into the existing
module nor the integration of the content element wizard is proven with
this commit.

See also https://wiki.typo3.org/Blueprints/WireframeComponent.