[FEATURE] Use TCA `columns` to configure multiple relationships
Instead of using a separate TCA configuration section for a relationship
between a content container and its content elements it seems to be much
cleaner to use columns and their corresponding TCA. Thus its also possible
to have multiple relationships with a single content container. Only the
user of the form data group has to know the column which has to be used.